Add `getDataGroupedPerPeriod` test
