+++ package.json
@@ -32,6 +32,7 @@
     "webpack-dev-server": "3.7.2"
   },
   "dependencies": {
+    "jquery": "^2.2.4",
     "moment": "2.24.0",
     "pinkie-promise": "2.0.1"
   }
