+++ src/main/js/utils/__mocks__/package-data.json
@@ -1,0 +1,98 @@
+{
+  "package-name": {
+    "2020-01-01": 1,
+    "2020-01-02": 13,
+    "2020-01-03": 0,
+    "2020-01-04": 0,
+    "2020-01-05": 6,
+    "2020-01-06": 26,
+    "2020-01-07": 26,
+    "2020-01-08": 17,
+    "2020-01-09": 21,
+    "2020-01-10": 12,
+    "2020-01-11": 1,
+    "2020-01-12": 5,
+    "2020-01-13": 18,
+    "2020-01-14": 32,
+    "2020-01-15": 19,
+    "2020-01-16": 13,
+    "2020-01-17": 21,
+    "2020-01-18": 4,
+    "2020-01-19": 5,
+    "2020-01-20": 19,
+    "2020-01-21": 41,
+    "2020-01-22": 17,
+    "2020-01-23": 19,
+    "2020-01-24": 33,
+    "2020-01-25": 0,
+    "2020-01-26": 0,
+    "2020-01-27": 25,
+    "2020-01-28": 16,
+    "2020-01-29": 18,
+    "2020-01-30": 17,
+    "2020-01-31": 11,
+    "2020-02-01": 1,
+    "2020-02-02": 4,
+    "2020-02-03": 21,
+    "2020-02-04": 23,
+    "2020-02-05": 12,
+    "2020-02-06": 23,
+    "2020-02-07": 18,
+    "2020-02-08": 2,
+    "2020-02-09": 5,
+    "2020-02-10": 26,
+    "2020-02-11": 31,
+    "2020-02-12": 22,
+    "2020-02-13": 34,
+    "2020-02-14": 16,
+    "2020-02-15": 3,
+    "2020-02-16": 10,
+    "2020-02-17": 14,
+    "2020-02-18": 24,
+    "2020-02-19": 10,
+    "2020-02-20": 17,
+    "2020-02-21": 24,
+    "2020-02-22": 2,
+    "2020-02-23": 5,
+    "2020-02-24": 31,
+    "2020-02-25": 27,
+    "2020-02-26": 46,
+    "2020-02-27": 35,
+    "2020-02-28": 14,
+    "2020-02-29": 4,
+    "2020-03-01": 5,
+    "2020-03-02": 26,
+    "2020-03-03": 16,
+    "2020-03-04": 23,
+    "2020-03-05": 25,
+    "2020-03-06": 18,
+    "2020-03-07": 0,
+    "2020-03-08": 5,
+    "2020-03-09": 18,
+    "2020-03-10": 15,
+    "2020-03-11": 9,
+    "2020-03-12": 24,
+    "2020-03-13": 70,
+    "2020-03-14": 2,
+    "2020-03-15": 8,
+    "2020-03-16": 21,
+    "2020-03-17": 28,
+    "2020-03-18": 29,
+    "2020-03-19": 34,
+    "2020-03-20": 63,
+    "2020-03-21": 4,
+    "2020-03-22": 44,
+    "2020-03-23": 42,
+    "2020-03-24": 28,
+    "2020-03-25": 32,
+    "2020-03-26": 18,
+    "2020-03-27": 23,
+    "2020-03-28": 7,
+    "2020-03-29": 8,
+    "2020-03-30": 37,
+    "2020-03-31": 29,
+    "2020-04-01": 27,
+    "2020-04-02": 40,
+    "2020-04-03": 21
+  }
+}
